reverted to basic make_forms script
